Create uniquechar(a).cpp
It's also okay to assume 256 characters. This would be the case in extended ASCII.  All the array elements are initially set to false. As we iterate over the string, set true at the index equal to the int value of the character. If at any time, we encounter that the array value is already true, it means the character with that int value is repeated.